RDKCOM-2912: REFPLTV-1146 Bring up 64bit support in Rpi4 (for both kernel an...
Reason for change: to match datatype compatibilty
Test Procedure: Build should be successful. bluetooth should build fine.
Risks: Low/None.
(cherry picked from commit f4f0446d18c78f4ac0aeed126c378943629f4a75)

+++ src/bt-ifce/btrCore_dbus_bluez5.c
@@ -2726,7 +2726,7 @@ btrCore_BTLEGattOps (
         stBTDeviceInfo  lstBTDeviceInfo;
         btrCore_BTGetDeviceInfo(apstlhBtIfce->pDBusConn, &lstBTDeviceInfo, objectPath);
         BTRCORELOG_DEBUG("device address is %s\n", lstBTDeviceInfo.pcAddress);
-        BTRCORELOG_DEBUG("Received data is %s with length %d\n", propertyValue, strlen(propertyValue));
+        BTRCORELOG_DEBUG("Received data is %s with length %lu\n", propertyValue, (unsigned long)strlen(propertyValue));
 
         /* Update data to the LE layer */
         if (apstlhBtIfce->fpcBTLeGattPath) {
@@ -4199,7 +4199,7 @@ BtrCore_BTSetProp (
     
     if (aenBtOpIfceType == enBTGattCharacteristic) {
         char* pVlaue = (char*)apvVal;
-        BTRCORELOG_INFO("Notified Value is %s with length %d\n", pVlaue, strlen(pVlaue));
+        BTRCORELOG_INFO("Notified Value is %s with length %lu\n", pVlaue, (unsigned long)strlen(pVlaue));
             
         DBusMessageIter lDBusMsgIter, signal_iter;
         DBusMessageIter lDBusMsgIterDictStr, lDBusMsgIterSubArray, lDBusMsgIterVariant;
